requirements.txt
pillow
scipy
torch>=1.9
torchvision>=0.10
tqdm>=4.41.0,<5
pandas>=1.1.5